API/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
aiohttp==3.9.5
pydantic==2.6.4
python-dotenv==1.0.1
openai==1.23.2
aiosqlite==0.19.0